clean up

+++ client/src/Components/ListItems/ListItems.js
@@ -3,30 +3,36 @@ import { InputGroup, FormControl } from 'react-bootstrap'
 import { MdDelete } from "react-icons/md";
 import './itemstyles.css'
 class ListItems extends Component {
-  state = {
-    done: this.props.itemData.done,
-    item: this.props.itemData.todo
-  }
+
   toggleStatus = async (e) => {
-    this.props.toggleStatus(this.props.itemData.index)
-    // console.log(this.state);
+    if (this.props.toggleStatus) {
+      this.props.toggleStatus(this.props.itemData.index)    
+    }
   }
-  deleteTodo=async()=>{
+  deleteTodo = async () => {
+    if (this.props.deleteTodo){
     this.props.deleteTodo(this.props.itemData.index)
+    }
   }
   render() {
-    
+
     return (
       <div>
-        
-        <li className="lists" key={this.props.itemData.index}>
-        <InputGroup className="mb-3">
-          <InputGroup.Prepend>
-            <InputGroup.Checkbox onClick={this.toggleStatus} />
-          </InputGroup.Prepend>
-          <FormControl aria-label="Text input with checkbox" value={this.props.itemData.todo} />
-          <MdDelete style={{width:'20px',height:'35px'}} onClick={this.deleteTodo}></MdDelete>
-        </InputGroup>
+
+        <li className="listStyle" key={this.props.itemData.index}>
+          <InputGroup className="mb-3">
+            <InputGroup.Prepend>
+              <InputGroup.Checkbox
+                onClick={this.toggleStatus} />
+            </InputGroup.Prepend>
+            <FormControl
+              
+              style={{ textDecoration: this.props.itemData.done ? 'line-through' : 'none' }}
+              defaultValue={this.props.itemData.todo} />
+            <MdDelete
+              style={{ width: '20px', height: '35px' }}
+              onClick={this.deleteTodo}></MdDelete>
+          </InputGroup>
         </li>
       </div>
     );
